requirements.txt
kafka-python==2.0.2
pyspark==3.4.2
cassandra-driver==3.29.0